backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pillow==10.4.0
pytesseract==0.3.13
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
numpy==1.26.4
pytest==8.3.3

